dashboard/requirements.txt
pytest
pylint
pytest--cov
streamlit
python-dotenv
boto3
altair
pandas
psycopg2